requirements.txt
transformers
pytorch
faiss
datasets
numpy
tqdm
seqeval
